feat(analytic): consolidate Clarity and Google Analytics setup into main module
